1. 전 커밋에서 long double로 바꿨던 변수들 double로 다시 되돌림 2. Indicator.hpp에 Engines/BarData.hpp를 직접 include 하여 지표 개발 편의성 증대 3. 기타 코드 정리

+++ Includes/Indicators/ExponentialMovingAverage.hpp
@@ -25,5 +25,5 @@ class ExponentialMovingAverage final : public Indicator {
   double alpha_;  // EMA 가중치
 
   void Initialize() override;
-  Numeric<long double> Calculate() override;
+  Numeric<double> Calculate() override;
 };
