+++ Includes/Indicators/Open.hpp
@@ -13,5 +13,5 @@ class Open final : public Indicator {
   int symbol_idx_;
 
   void Initialize() override;
-  Numeric<long double> Calculate() override;
+  Numeric<double> Calculate() override;
 };
