+++ Sources/Cpp/Indicators/Open.cpp
@@ -1,11 +1,6 @@
 // 파일 헤더
 #include "Indicators/Open.hpp"
 
-// 내부 헤더
-#include "Engines/BarData.hpp"
-#include "Engines/BarHandler.hpp"
-#include "Engines/BaseBarHandler.hpp"
-
 Open::Open(const string& name, const string& timeframe, const Plot& plot)
     : Indicator(name, timeframe, plot), symbol_idx_(-1) {}
 
@@ -14,6 +9,6 @@ void Open::Initialize() {
   symbol_idx_ = bar_->GetCurrentSymbolIndex();
 }
 
-Numeric<long double> Open::Calculate() {
+Numeric<double> Open::Calculate() {
   return reference_bar_->GetBar(symbol_idx_, bar_->GetCurrentBarIndex()).open;
 }
